arreglo ruta

+++ BD/script_creacion_inicial.sql
@@ -2759,7 +2759,7 @@ GO
 
 -------------------------------Servicios de una ruta-------------------------------
 CREATE FUNCTION [ABSTRACCIONX4].ServiciosDeRuta
-	(@IdRuta TINYINT)
+	(@IdRuta SMALLINT)
 RETURNS @Servicios TABLE (tipoServicio VARCHAR(30))
 AS
 BEGIN
